gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
web3==7.6.0
httpx==0.27.0
pyjwt==2.9.0
pydantic==2.9.0
pydantic-settings==2.5.0
structlog==24.4.0
eth-account==0.13.4
asyncpg>=0.29.0
